+++ src/main/java/com/inti/services/interfaces/IChauffeurService.java
@@ -7,8 +7,8 @@ import com.inti.entities.Chauffeur;
 public interface IChauffeurService {
 	
 	public List<Chauffeur> findAll();
-	public Chauffeur findById(int idChauffeur);
+	public Chauffeur findById(Long idChauffeur);
 	public void save(Chauffeur chauffeur);
-	public void deleteById(int idChauffeur);
+	public void deleteById(Long idChauffeur);
 
 }
