+++ src/main/java/com/inti/services/interfaces/IMoyenTransportService.java
@@ -7,8 +7,8 @@ import com.inti.entities.MoyenTransport;
 public interface IMoyenTransportService {
 	
 	public List<MoyenTransport> findAll();
-	public MoyenTransport findById(int idMoyenTransport);
+	public MoyenTransport findById(Long idMoyenTransport);
 	public void save(MoyenTransport moyenTransport);
-	public void deleteById(int idMoyenTransport);
+	public void deleteById(Long idMoyenTransport);
 
 }
